Add remove() function to remove a key from the HashMap
self.remove() finds the position of the intended key and does a
swap_remove. The swap_remove moves the key from the intended position
to the last position of the vector and swaps the empty position
with the last key of the vector. Finally, the last position of the
vector is truncated.